+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [1.2.0] - 2025-12-01
+
+### Added
+
+- New implementation of countries
+
+### Changed
+
+- Update implementation of some countries
+
 ## [1.1.0] - 2023-07-18
 
 ### Added